requirements.txt
boto3==1.34.79
textract==1.6.5
langchain==0.1.14
openai==0.28
python-dotenv==1.0.1
pymilvus==2.3.5
google-serp-api==1.0.4
PYPDF2==3.0.1
google-search-results==2.4.2
